+++ src/main/java/ca/bcit/comp2522/termproject/comp2522202330termprojectmartincharliegame/Item.java
@@ -1,5 +1,5 @@
 package ca.bcit.comp2522.termproject.comp2522202330termprojectmartincharliegame;
 
 public interface Item {
-    public double getValue();
+    public int getValue();
 }
